Fix type

+++ src/main/java/com/hmh/mmp/dto/debit/DebitPagingDTO.java
@@ -1,14 +1,11 @@
 package com.hmh.mmp.dto.debit;
 
-import com.hmh.mmp.entity.BankEntity;
 import lombok.Data;
-
-import java.time.LocalDate;
 
 @Data
 public class DebitPagingDTO extends DebitSaveDTO {
     private Long debitId;
 
-    public DebitPagingDTO(Long id, Double debitGet, String debitMemo, String debitName, Long minusAsset, String account, String debitPhotoName) {
+    public DebitPagingDTO(Long id, Double debitGet, String debitMemo, String debitName, Double minusAsset, String account, String debitPhotoName) {
     }
 }
